Naive bayes implementation
Analysis code is in analysis.py
full-corpus-1.csv is the tweet database. Apple, google, microsoft, are the training tweets, twitter is the test.
data required to learn is stored in word_data.txt (word appearance counts in pos, neg, neutral tweets).
tweet_spread.txt is a record of types of tweets read.
report.txt stores a record of tweets that were categorized wrongly.